Change all the type of double to unsigned long
note that I use a typedef to define unsigned long as UINT

+++ test.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include <stdlib.h>
 #include "ds.h"
 #include "connect.h"
 
@@ -47,7 +48,7 @@ int main(int argc, char * argv[]){
 	// (blockage.num - 2) denotes the index of first point
 	dijkstra(&blockage,g_size-2);
 
-	printf("s to t: %lf\n",g[g_size-2][g_size-1]);
+	printf("s to t: %ld\n",(int)g[g_size-2][g_size-1]);
 
 	free_all();
 	return 0;
